edit cocktail and separate stores
